Create custom styling for map title, container and pop-up container
To ensure that there is enough room for the full amount of map information, moved the map to 2/3 of the browser, leaving the remaining for the popup content to appear on click.

Add font styles from Google Fonts - specifically with exploring fonts that allows moph and shadow length customization (very, very COOL!).