LAB 1-2: Add new spatiotemporal data set
This data set more accurately meets the lab requirement of not only finding a spatiotemportal data set with at minimum 15 locations and 7 time frames, but also is an appropriate for mapping with proportional symbology. The data was provided by the Portland Police Bureau https://www.portland.gov/police/open-data/crime-statistics#toc-dashboard-walkthrough  and downloaded by year from 2015 to 2023. After which I performed a combination of spatial data transformation, geoprocessing, and data wrangling tasks to prepare your CSV datasets for visualization as a GeoJSON layer on a map.